+++ r/R/port4me.R
@@ -28,39 +28,63 @@ parse_ports <- function(ports) {
   ports <- paste(ports, collapse = ",")
   ports <- gsub(" ", "", ports, fixed = TRUE)
   ports <- unlist(strsplit(ports, split = ","))
-  bad <- grep("^[[:digit:]]+$", ports, invert = TRUE, value = TRUE)
+  bad <- grep("^([[:digit:]]+|[[:digit:]]+-[[:digit:]]+)$", ports, invert = TRUE, value = TRUE)
   if (length(bad) > 0) {
     stop(sprintf("Syntax error in port specification: %s", spec))
   }
-  ports <- as.integer(ports)
+  ports <- lapply(ports, FUN = function(spec) {
+    pattern <- "^([[:digit:]]+)-([[:digit:]]+)$"
+    if (grepl(pattern, spec)) {
+      from <- as.integer(gsub(pattern, "\\1", spec))
+      to <- as.integer(gsub(pattern, "\\2", spec))
+      from:to
+    } else {
+      as.integer(spec)
+    }
+  })
+  ports <- unlist(ports, use.names = FALSE)
   stopifnot(!anyNA(ports))
+  if (is.null(ports)) ports <- integer(0L)
   ports
 }
 
-port4me_include <- function() {
-  include <- NULL
+port4me_prepend <- function() {
+  ports <- NULL
 
-  for (name in c("PORT4ME_INCLUDE", "PORT4ME_INCLUDE_SITE")) {
+  for (name in c("PORT4ME_PREPEND", "PORT4ME_PREPEND_SITE")) {
     arg <- Sys.getenv(name, "")
     ports <- parse_ports(arg)
-    include <- c(include, ports)
+    ports <- c(ports, ports)
   }
-  include <- unique(include)
-  
-  include
+  ports <- unique(ports)
+
+  ports
 }
 
 port4me_exclude <- function() {
-  exclude <- NULL
+  ports <- NULL
 
   for (name in c("PORT4ME_EXCLUDE", "PORT4ME_EXCLUDE_SITE")) {
     arg <- Sys.getenv(name, "")
     ports <- parse_ports(arg)
-    exclude <- c(exclude, ports)
+    ports <- c(ports, ports)
   }
-  exclude <- unique(exclude)
+  ports <- unique(ports)
   
-  exclude
+  ports
+}
+
+port4me_include <- function() {
+  ports <- NULL
+
+  for (name in c("PORT4ME_INCLUDE", "PORT4ME_INCLUDE_SITE")) {
+    arg <- Sys.getenv(name, "")
+    ports <- parse_ports(arg)
+    ports <- c(ports, ports)
+  }
+  ports <- unique(ports)
+  
+  ports
 }
 
 port4me_skip <- function() {
@@ -69,7 +93,7 @@ port4me_skip <- function() {
   skip
 }
 
-port4me <- function(user = port4me_user(), tool = port4me_tool(), include = port4me_include(), exclude = port4me_exclude(), skip = port4me_skip(), list = NULL, test = NULL, max_tries = 1000L, must_work = TRUE) {
+port4me <- function(user = port4me_user(), tool = port4me_tool(), prepend = port4me_prepend(), include = port4me_exclude(), exclude = port4me_exclude(), skip = port4me_skip(), list = NULL, test = NULL, max_tries = 1000L, must_work = TRUE) {
   stopifnot(length(user) == 1L, is.character(user), !is.na(user))
   stopifnot(is.null(tool) || is.character(tool), !anyNA(tool))
   if (!is.null(list)) {
@@ -77,10 +101,12 @@ port4me <- function(user = port4me_user(), tool = port4me_tool(), include = port
   }
   stopifnot(length(max_tries) == 1L, is.numeric(max_tries), !is.na(max_tries), max_tries > 0, is.finite(max_tries))
   max_tries <- as.integer(max_tries)
-  if (is.character(include)) include <- parse_ports(include)
-  stopifnot(is.numeric(include), !anyNA(include), all(include > 0), all(include <= 65535))
+  if (is.character(prepend)) prepend <- parse_ports(prepend)
+  stopifnot(is.numeric(prepend), !anyNA(prepend), all(prepend > 0), all(prepend <= 65535))
   if (is.character(exclude)) exclude <- parse_ports(exclude)
   stopifnot(is.numeric(exclude), !anyNA(exclude), all(exclude > 0), all(exclude <= 65535))
+  if (is.character(include)) include <- parse_ports(include)
+  stopifnot(is.numeric(include), !anyNA(include), all(include > 0), all(include <= 65535))
   stopifnot(length(skip) == 1L, is.numeric(skip), !is.na(skip), skip >= 0, is.finite(skip), skip < max_tries)
   skip <- as.integer(skip)
   if (!is.null(test)) {
@@ -101,13 +127,14 @@ port4me <- function(user = port4me_user(), tool = port4me_tool(), include = port
   ports <- integer(0)
   count <- 0L
   while (count <= max_tries) {
-    if (length(include) > 0) {
-      port <- include[1]
-      include <- include[-1]
+    if (length(prepend) > 0) {
+      port <- prepend[1]
+      prepend <- prepend[-1]
     } else {
       port <- lcg_port()
     }
     if (port %in% exclude) next
+    if (length(include) > 0 && (! port %in% include)) next
     count <- count + 1L
     if (count <= skip) next
     if (is.null(list)) {